Struct rtic_monotonics::TimerQueue

pub struct TimerQueue<Mono>where
    Mono: Monotonic,{ /* private fields */ }

A generic timer queue for async executors.

Blocking

The internal priority queue uses global critical sections to manage access. This means that awaiting a delay will cause a lock of the entire system for O(n) time. In practice the lock duration is ~10 clock cycles per element in the queue.

Safety

This timer queue is based on an intrusive linked list, and by extension the links are strored on the async stacks of callers. The links are deallocated on drop or when the wait is complete.

Do not call mem::forget on an awaited future, or there will be dragons!

Implementations

impl<Mono> TimerQueue<Mono>where Mono: Monotonic,

pub const fn new() -> TimerQueue<Mono>

Make a new queue.

pub fn now(&self) -> <Mono as Monotonic>::Instant

Forwards the Monotonic::now() method.

pub fn initialize(&self, monotonic: Mono)

Takes the initialized monotonic to initialize the TimerQueue.

pub unsafe fn on_monotonic_interrupt(&self)

Call this in the interrupt handler of the hardware timer supporting the Monotonic

Safety

It’s always safe to call, but it must only be called from the interrupt of the monotonic timer for correct operation.

pub async fn timeout_at<F>( &self, instant: <Mono as Monotonic>::Instant, future: F ) -> Result<<F as Future>::Output, TimeoutError>where F: Future,

Timeout at a specific time.

pub async fn timeout_after<F>( &self, duration: <Mono as Monotonic>::Duration, future: F ) -> Result<<F as Future>::Output, TimeoutError>where F: Future,

Timeout after at least a specific duration.

pub async fn delay(&self, duration: <Mono as Monotonic>::Duration)

Delay for at least some duration of time.

pub async fn delay_until(&self, instant: <Mono as Monotonic>::Instant)

Delay to some specific time instant.