Trait rtic_monotonics::TimerQueueBasedMonotonic

pub trait TimerQueueBasedMonotonic {
    type Backend: TimerQueueBackend;
    type Instant: TimerQueueBasedInstant<Ticks = <Self::Backend as TimerQueueBackend>::Ticks> + Add<Self::Duration, Output = Self::Instant> + Sub<Self::Duration, Output = Self::Instant, Output = Self::Duration> + Sub;
    type Duration: TimerQueueBasedDuration<Ticks = <Self::Backend as TimerQueueBackend>::Ticks>;
}

A Monotonic that is backed by the TimerQueue.

Required Associated Types

type Backend: TimerQueueBackend

The backend for the timer queue

type Instant: TimerQueueBasedInstant<Ticks = <Self::Backend as TimerQueueBackend>::Ticks> + Add<Self::Duration, Output = Self::Instant> + Sub<Self::Duration, Output = Self::Instant, Output = Self::Duration> + Sub

The type for instant, defining an instant in time.

Note: In all APIs in RTIC that use instants from this monotonic, this type will be used.

type Duration: TimerQueueBasedDuration<Ticks = <Self::Backend as TimerQueueBackend>::Ticks>

The type for duration, defining a duration of time.

Note: In all APIs in RTIC that use duration from this monotonic, this type will be used.

Object Safety

This trait is not object safe.