Crate rtic_monotonics

In-tree implementations of the rtic_time::Monotonic (reexported) trait for timers & clocks found on commonly used microcontrollers.

To enable the implementations, you must enable a feature for the specific MCU you’re targeting.

Cortex-M Systick

The systick monotonic works on all cortex-M parts, and requires that the feature cortex-m-systick is enabled.

RP2040

The RP2040 monotonics require that the rp2040 feature is enabled.

i.MX RT

The i.MX RT monotonics require that the feature imxrt_gpt1 or imxrt_gpt2 is enabled.

nRF

nRF monotonics require that one of the available nrf52* features is enabled.

All implementations of timers for the nRF52 family are documented here. Monotonics that are not available on all parts in this family will have an Available on crate features X only tag, describing what parts do support that monotonic. Monotonics without an Available on crate features X only tag are available on any nrf52* feature.

Structs

TimeoutError

This indicates that there was a timeout.

TimerQueue

A generic timer queue for async executors.

Traits

InterruptToken

This marker is implemented on an interrupt token to enforce that the right tokens are given to the correct monotonic implementation.

Monotonic

A monotonic clock / counter definition.