Trait rtic_monotonics::Monotonic

pub trait Monotonic {
    type Instant: Ord + Copy + Add<Self::Duration, Output = Self::Instant> + Sub<Self::Duration, Output = Self::Instant, Output = Self::Duration> + Sub;
    type Duration;

    const ZERO: Self::Instant;
    const TICK_PERIOD: Self::Duration;

    // Required methods
    fn now() -> Self::Instant;
    fn set_compare(instant: Self::Instant);
    fn clear_compare_flag();
    fn pend_interrupt();

    // Provided methods
    fn should_dequeue_check(release_at: Self::Instant) -> bool { ... }
    fn on_interrupt() { ... }
    fn enable_timer() { ... }
    fn disable_timer() { ... }
}

A monotonic clock / counter definition.

Correctness

The trait enforces that proper time-math is implemented between Instant and Duration. This is a requirement on the time library that the user chooses to use.

Required Associated Types

type Instant: Ord + Copy + Add<Self::Duration, Output = Self::Instant> + Sub<Self::Duration, Output = Self::Instant, Output = Self::Duration> + Sub

The type for instant, defining an instant in time.

Note: In all APIs in RTIC that use instants from this monotonic, this type will be used.

type Duration

The type for duration, defining an duration of time.

Note: In all APIs in RTIC that use duration from this monotonic, this type will be used.

Required Associated Constants

const ZERO: Self::Instant

The time at time zero.

const TICK_PERIOD: Self::Duration

The duration between two timer ticks.

Required Methods

fn now() -> Self::Instant

Get the current time.

fn set_compare(instant: Self::Instant)

Set the compare value of the timer interrupt.

Note: This method does not need to handle race conditions of the monotonic, the timer queue in RTIC checks this.

fn clear_compare_flag()

Clear the compare interrupt flag.

fn pend_interrupt()

Pend the timer’s interrupt.

Provided Methods

fn should_dequeue_check(release_at: Self::Instant) -> bool

Override for the dequeue check, override with timers that have bugs.

E.g. nRF52 RTCs needs to be dequeued if the time is within 4 ticks.

fn on_interrupt()

Optional. Runs on interrupt before any timer queue handling.

fn enable_timer()

Optional. This is used to save power, this is called when the timer queue is not empty.

Enabling and disabling the monotonic needs to propagate to now so that an instant based of now() is still valid.

NOTE: This may be called more than once.

fn disable_timer()

Optional. This is used to save power, this is called when the timer queue is empty.

Enabling and disabling the monotonic needs to propagate to now so that an instant based of now() is still valid.

NOTE: This may be called more than once.

Object Safety

This trait is not object safe.